added funcion to deleting product from cart, transferred some actions from categories slice to cart slice
