+++ src/store/store.ts
@@ -1,9 +1,11 @@
 import { configureStore } from '@reduxjs/toolkit'
 import categoriesReducer from './slices/categoriesSlice'
+import cartReducer from './slices/cartSlice'
 
 export const store = configureStore({
   reducer: {
     categories: categoriesReducer,
+    user: cartReducer
   },
 })
